+++ src/LightBulbs/Providers/IntegrityLightBulbProvider.cs
@@ -1,0 +1,40 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.Composition;
+using Microsoft.Html.Core.Tree.Nodes;
+using Microsoft.Html.Core.Tree.Utility;
+using Microsoft.Html.Editor.SuggestedActions;
+using Microsoft.VisualStudio.Language.Intellisense;
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.Utilities;
+using Microsoft.Web.Core.ContentTypes;
+
+namespace HtmlTools
+{
+    [Export(typeof(IHtmlSuggestedActionProvider))]
+    [ContentType(HtmlContentTypeDefinition.HtmlContentType)]
+    [Name("Calculate Integrity Light Bulb Provider")]
+    class IntegrityLightBulbProvider : IHtmlSuggestedActionProvider
+    {
+        public IEnumerable<ISuggestedAction> GetSuggestedActions(ITextView textView, ITextBuffer textBuffer, int caretPosition, ElementNode element, AttributeNode attribute, HtmlPositionType positionType)
+        {
+            return new ISuggestedAction[] {
+                new IntegrityLightBulbAction(textView, textBuffer, element)
+            };
+        }
+
+        public bool HasSuggestedActions(ITextView textView, ITextBuffer textBuffer, int caretPosition, ElementNode element, AttributeNode attribute, HtmlPositionType positionType)
+        {
+            if (!element.StartTag.Contains(caretPosition))
+                return false;
+
+            string url = (element.GetAttribute("src") ?? element.GetAttribute("href"))?.Value;
+
+            if (string.IsNullOrEmpty(url) || (!url.Contains("://") && !url.StartsWith("//")))
+                return false;
+
+            return element.IsElement("style") || element.IsElement("script");
+        }
+    }
+}