+++ src/LightBulbs/Providers/StyleScriptLightBulbProvider.cs
@@ -1,0 +1,37 @@
+﻿using System.Collections.Generic;
+using System.ComponentModel.Composition;
+using Microsoft.Html.Core.Tree.Nodes;
+using Microsoft.Html.Core.Tree.Utility;
+using Microsoft.Html.Editor.SuggestedActions;
+using Microsoft.VisualStudio.Language.Intellisense;
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.Utilities;
+using Microsoft.Web.Core.ContentTypes;
+
+namespace HtmlTools
+{
+    [Export(typeof(IHtmlSuggestedActionProvider))]
+    [ContentType(HtmlContentTypeDefinition.HtmlContentType)]
+    [Name("HtmlStyleScriptLightBulbProvider")]
+    class HtmlStyleScriptLightBulbProvider : IHtmlSuggestedActionProvider
+    {
+        public IEnumerable<ISuggestedAction> GetSuggestedActions(ITextView textView, ITextBuffer textBuffer, int caretPosition, ElementNode element, AttributeNode attribute, HtmlPositionType positionType)
+        {
+            return new ISuggestedAction[] {
+                new HtmlExtractLightBulbAction(textView, textBuffer, element),
+            };
+        }
+
+        public bool HasSuggestedActions(ITextView textView, ITextBuffer textBuffer, int caretPosition, ElementNode element, AttributeNode attribute, HtmlPositionType positionType)
+        {
+            if (!element.StartTag.Contains(caretPosition))
+                return false;
+
+            if (element.InnerRange.Length < 5)
+                return false;
+
+            return element.IsStyleBlock() || element.IsJavaScriptBlock();
+        }
+    }
+}